refactor: LoginPage 에서 ID 검증 로직 추상화
